% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{c_data}
\alias{c_data}
\title{Combine results of run_geobam}
\usage{
c_data(result1, result2)
}
\arguments{
\item{result1}{data.frame}

\item{result2}{data.frame}
}
\value{
data.frame
}
\description{
Combine function for foreach statement.
}
