% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{check_observations}
\alias{check_observations}
\title{Checks if observation data is valid.}
\usage{
check_observations(width, d_x_area, slope2, qhat)
}
\arguments{
\item{width}{matrix}

\item{d_x_area}{matrix}

\item{slope2}{matrix}

\item{qhat}{vector}
}
\value{
list of valid observations or an empty list if there are none
}
\description{
Checks if observation data is valid.
}
