% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{concatenate_invalid}
\alias{concatenate_invalid}
\title{Insert NA values back into posterior list vectors to account for invalid
nodes}
\usage{
concatenate_invalid(posterior_list, invalid_nodes, invalid_time)
}
\arguments{
\item{posterior_list}{list of posterior data vectors}

\item{invalid_nodes}{list of invalid node indexes}

\item{invalid_time}{list of invalid time indexes}
}
\value{
list of posterior data with NA in place of invalid nodes
}
\description{
Insert NA values back into posterior list vectors to account for invalid
nodes
}
