% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{create_dimensions}
\alias{create_dimensions}
\title{Create netCDF dimensions}
\usage{
create_dimensions(nx, nt)
}
\arguments{
\item{nx}{integer vector that represents number of nodes}

\item{nt}{integer vector that represents time steps}
}
\value{
list ncdim
}
\description{
Create netCDF dimensions
}
