% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{create_nc_file}
\alias{create_nc_file}
\title{Create netCDF output file}
\usage{
create_nc_file(var_list, output_dir)
}
\arguments{
\item{var_list}{list ncvar}

\item{output_dir}{character}
}
\value{
ncdf4
}
\description{
Create netCDF output file
}
