% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{create_posterior_list}
\alias{create_posterior_list}
\title{Creates data.frame of NA values for each geoBAM posterior.}
\usage{
create_posterior_list(reachid, nx, nt)
}
\arguments{
\item{reachid}{string reach identifier}

\item{nx}{integer number of nodes}

\item{nt}{integer number of time steps}
}
\value{
list of posteriors set to NA
}
\description{
Creates data.frame of NA values for each geoBAM posterior.
}
