% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{create_vars}
\alias{create_vars}
\title{Create a list of netcdf4 variables}
\usage{
create_vars(reachid, nt_dim, nx_dim)
}
\arguments{
\item{reachid}{string reach identifier}

\item{nt_dim}{ncdim}

\item{nx_dim}{ncdim}
}
\value{
list ncvar
}
\description{
Create list of mean and sigma netcdf variables grouped by reachid.
}
