% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{extract_geobam_posteriors}
\alias{extract_geobam_posteriors}
\title{Title}
\usage{
extract_geobam_posteriors(geobam_data, priors)
}
\arguments{
\item{geobam_data}{object that contains geobam observation data}

\item{priors}{object that contains prior data}
}
\value{
list of mean and sd values for the chain
}
\description{
Title
}
