% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_input}
\alias{get_input}
\title{Get netcdf data from file}
\usage{
get_input(file)
}
\arguments{
\item{file}{netcdf}
}
\value{
list matrix
}
\description{
Retrieve geoBAM input variables and replace fill values with NA.
}
