% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_invalid}
\alias{get_invalid}
\title{Checks if observation parameter has valid nx (nodes) and valid nt (time steps)}
\usage{
get_invalid(obs)
}
\arguments{
\item{obs}{matrix}
}
\value{
list of invalid nodes and invalid time steps
}
\description{
Checks if observation parameter has valid nx (nodes) and valid nt (time steps)
}
