% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{get_posteriors}
\alias{get_posteriors}
\title{Generate and extract geobam posteriors from reach data.}
\usage{
get_posteriors(reach_data)
}
\arguments{
\item{reach_data}{list of reach data (input observations)}
}
\value{
list of mean and sd posteriors
}
\description{
Generate and extract geobam posteriors from reach data.
}
