% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{get_reach_id}
\alias{get_reach_id}
\title{Retrieve reach id from SWOT file name.}
\usage{
get_reach_id(swot_file)
}
\arguments{
\item{swot_file}{Path to SWOT file}
}
\value{
String reach identifier
}
\description{
Retrieve reach id from SWOT file name.
}
