% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{get_sd}
\alias{get_sd}
\title{Column-wise standard deviation}
\usage{
get_sd(posterior)
}
\arguments{
\item{posterior}{list}
}
\value{
list of standard deviations
}
\description{
Get the column-wise standard deviation
}
