% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{parse_reach_data}
\alias{parse_reach_data}
\title{Get netcdf data from files}
\usage{
parse_reach_data(swot_file, sword_file)
}
\arguments{
\item{swot_file}{netcdf}

\item{sword_file}{netcdf}
}
\value{
list matrix
}
\description{
Retrieve geoBAM input variables and replace fill values with NA.
}
