% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Execute geobam on data}
\usage{
process_data(
  input_dir,
  output_dir,
  env,
  cores = NA,
  partition = NA,
  max_jobs = NA,
  as_job_array = FALSE
)
}
\arguments{
\item{input_dir}{character vecto}

\item{output_dir}{character}

\item{env}{character name of environment to execute in (i.e. "workstation" or "slurm")}

\item{cores}{number of cores to run on (for workstation run only)}

\item{partition}{string partition name to run jobs on (slurm run only)}

\item{max_jobs}{integer max number of jobs to be run concurrently (slurm run only)}

\item{as_job_array}{boolean indication to run execute as job array (slurm run only)}
}
\value{
None
}
\description{
Execute geobam on data in user-provided input directory and outputs a single
netCDF file of results organized as variables grouped by reachid to
user-provided output directory. Directories need to be passed in as absolute
paths as character values.
}
