% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{run_geobam}
\alias{run_geobam}
\title{Run geobam}
\usage{
run_geobam(reachid, data_dir, output_dir)
}
\arguments{
\item{reachid}{string reach identifier}

\item{data_dir}{directory of SWOT and SoS input files}

\item{output_dir}{directory where posterior output will be written}
}
\description{
Write output from geoBAM execution on each reach data input.
}
