% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{run_hpc}
\alias{run_hpc}
\title{Run geoBAM in parallel across nodes on specified number of cores.}
\usage{
run_hpc(reaches, input_dir, output_dir, nodes, cpus_per_node)
}
\arguments{
\item{reaches}{list of string reach identifiers}

\item{input_dir}{string input directory}

\item{output_dir}{string output directory}

\item{nodes}{integer}

\item{cpus_per_node}{integer}
}
\description{
Run geoBAM in parallel across nodes on specified number of cores.
}
