% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{run_mpi}
\alias{run_mpi}
\title{Run geoBAM in parallel on multiple processes across nodes using doMPI backend
to foreach function.}
\usage{
run_mpi(reaches, input_dir, output_dir)
}
\arguments{
\item{reaches}{list of string reach identifiers}

\item{input_dir}{string input directory}

\item{output_dir}{string output directory}
}
\description{
Run geoBAM in parallel on multiple processes across nodes using doMPI backend
to foreach function.
}
