% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{run_slurm}
\alias{run_slurm}
\title{Run geoBAM in parallel using batchtools where each reach is submitted
as a job.}
\usage{
run_slurm(reaches, input_dir, output_dir, ncpus)
}
\arguments{
\item{reaches}{list of string reach identifiers}

\item{input_dir}{string input directory}

\item{output_dir}{string output directory}

\item{ncpus}{integer number of cpus per job.}
}
\description{
Run geoBAM in parallel using batchtools where each reach is submitted
as a job.
}
