% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{run_workstation}
\alias{run_workstation}
\title{Run geoBAM in parallel on multiple processes on single workstation.}
\usage{
run_workstation(
  reaches,
  input_dir,
  output_dir,
  cores = parallel::detectCores()
)
}
\arguments{
\item{reaches}{list of string reach identifiers}

\item{input_dir}{string input directory}

\item{output_dir}{string output directory}

\item{cores}{integer number of cores in system (default is determed by detectCores)}
}
\description{
Run geoBAM in parallel on multiple processes on single workstation.
}
