% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{update_posteriors}
\alias{update_posteriors}
\title{Update posterior list with mean and sd posteriors in posteriors parameter}
\usage{
update_posteriors(posterior_list, posteriors)
}
\arguments{
\item{posterior_list}{list of posterior data}

\item{posteriors}{list of posterior mean and sd values}
}
\value{
data.frame of updated posterior data
}
\description{
Update posterior list with mean and sd posteriors in posteriors parameter
}
