% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{write_netcdf}
\alias{write_netcdf}
\title{Write netCDF}
\usage{
write_netcdf(result_df, output_dir)
}
\arguments{
\item{result_df}{data.frame}

\item{output_dir}{character}
}
\value{
None
}
\description{
Writes netcdf for results dataframe to output directory as a single file
called 'geobamr_output_ALL.nc' with variables groups by Reach ID.
}
