% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{write_vars}
\alias{write_vars}
\title{Write netcdf variables}
\usage{
write_vars(df_slice, var_list, nc_out)
}
\arguments{
\item{df_slice}{data.frame}

\item{var_list}{list ncvar}

\item{nc_out}{ncdf4}
}
\value{
None
}
\description{
Write mean and sigma variables to netcdf file.
}
