% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{create_dimensions}
\alias{create_dimensions}
\title{Create netCDF dimensions}
\usage{
create_dimensions(df_slices)
}
\arguments{
\item{df_slices}{data.frame}
}
\value{
list ncdim
}
\description{
Create netCDF dimensions
}
