% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_input_data}
\alias{get_input_data}
\title{Get netcdf data from files}
\usage{
get_input_data(input_dir)
}
\arguments{
\item{input_dir}{file directory}
}
\value{
list matrix
}
\description{
Retrieve geoBAM input variables and replace fill values with NA.
}
