% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Execute geobam on data}
\usage{
process_data(input_dir, output_dir)
}
\arguments{
\item{input_dir}{character}

\item{output_dir}{character}
}
\value{
None
}
\description{
Execute geobam on data in user-provided input directory and outputs a single
netCDF file of results organized as variables grouped by reachid to
user-provided output directory. Directories need to be passed in as absolute
paths as character values.
}
