% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Execute geobam on data}
\usage{
process_data(input_dir, output_dir, env, cores = NA, ncpus = NA)
}
\arguments{
\item{input_dir}{character vecto}

\item{output_dir}{character}

\item{env}{character name of environment to execute in (i.e. "workstation" or "slurm")}

\item{cores}{number of cores to run on (for workstation run only)}

\item{ncpus}{number of cpus required per job (for HPC SLURM run only)}
}
\value{
None
}
\description{
Execute geobam on data in user-provided input directory and outputs a single
netCDF file of results organized as variables grouped by reachid to
user-provided output directory. Directories need to be passed in as absolute
paths as character values.
}
