% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{run_geobam}
\alias{run_geobam}
\title{Run geobam}
\usage{
run_geobam(reach)
}
\arguments{
\item{reach}{list of reach data}
}
\value{
data.frame
}
\description{
Get output from geoBAMr execution on each reach data input.
}
