% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobam_data.R
\name{update_posteriors}
\alias{update_posteriors}
\title{Update posterior list with mean and sd posteriors in posteriors parameter}
\usage{
update_posteriors(index, posteriors, posterior_list)
}
\arguments{
\item{index}{integer number index (chain number) for both parameter lists}

\item{posteriors}{list of posterior mean and sd values}

\item{posterior_list}{list of posterior data}
}
\value{
list of updated posterior data for chain
}
\description{
Update posterior list with mean and sd posteriors in posteriors parameter
}
