% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{write_netcdf}
\alias{write_netcdf}
\title{Write netCDF}
\usage{
write_netcdf(reach_data, posterior_list, output_dir)
}
\arguments{
\item{reach_data}{list of reach data (observations)}

\item{posterior_list}{list of posterior data}

\item{output_dir}{string path to output directory}
}
\value{
None
}
\description{
Writes netCDF of posterior data to output directory labelled with reach
identifier.
}
