% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{write_vars}
\alias{write_vars}
\title{Write netcdf variables}
\usage{
write_vars(index, posterior_list, var_list, nc_out)
}
\arguments{
\item{index}{integer index number for parameter lists}

\item{posterior_list}{list of posterior data}

\item{var_list}{list of ncvar}

\item{nc_out}{ncdf4}
}
\value{
None
}
\description{
Write reachid, mean and sigma variables to netcdf file. Also includes
a valid global attribute to indicate validity of reach data.
}
